Third-party-img-read/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pillow==10.0.0
python-dotenv==1.0.0
pytesseract==0.3.10
numpy==1.25.2
pandas==2.0.3